+++ src/modules/biascorrectionConv_mod.f90
@@ -30,7 +30,7 @@ MODULE biasCorrectionConv_mod
   implicit none
   save
   private
-  public               :: bcc_readConfig, bcc_applyAIBcor, bcc_applyGPBcor
+  public               :: bcc_applyAIBcor, bcc_applyGPBcor
   public               :: bcc_biasActive
 
   integer, parameter :: nPhases=3, nLevels=5, nAircraftMax=100000 
@@ -44,11 +44,12 @@ MODULE biasCorrectionConv_mod
 
 
   integer  :: nbAircrafts, nbGpStations
-  real, allocatable  :: ttCorrections(:,:,:)
-  real, allocatable  :: ztdCorrections(:)
+  real(8), allocatable  :: ttCorrections(:,:,:)
+  real(8), allocatable  :: ztdCorrections(:)
   character(len=9), allocatable :: aircraftIds(:), gpsStations(:)
   
   logical :: bcc_aiBiasActive, bcc_gpBiasActive
+  logical :: initialized = .false.
   
   ! Bias correction files (must be in program working directory)
   character(len=8), parameter :: aiBcFile = "ai_bcors", gpBcFile = "gp_bcors"
@@ -95,42 +96,46 @@ CONTAINS
     bcc_aiBiasActive = aiBiasActive
     bcc_gpBiasActive = gpBiasActive
     
+    initialized = .true.
+    
     
   end subroutine bcc_readConfig
 
   !-----------------------------------------------------------------------
-  ! bcc_readAIBcor
+  ! bcc_readAIBiases
   !-----------------------------------------------------------------------
-  subroutine bcc_readAIBcor(biasCorrectionFileName)
+  subroutine bcc_readAIBiases(biasEstimateFile)
     !
-    ! :Purpose: Read TT bias corrections. 
+    ! :Purpose: Read aircraft (AI) TT bias estimates from bias file and fill bias correction array ttCorrections. 
     !           The first line of the file is the number of aircraft plus one.
     !           The rest of the file gives 15 values of Mean O-A for each aircraft, with each (AC,value) line written in format "a9,1x,f6.2".
     !           The order is the same as what is written by genbiascorr script genbc.aircraft_bcor.py.
-    !           The first "aircraft" (AC name = BULKBCORS) values are the bulk corrections by layer for All-AC (first 5 values), AIREP/ADS 
+    !           The first "aircraft" (AC name = BULKBCORS) values are the bulk biases by layer for All-AC (first 5 values), AIREP/ADS 
     !           (second 5 values) and AMDAR/BUFR (last 5 values).
     !           Missing value = 99.0.
     !
     implicit none
     !Arguments:
-    character(len=*), intent(in) :: biasCorrectionFileName
+    character(len=*), intent(in) :: biasEstimateFile
 
     !Locals:
     integer :: ierr, nulcoeff
     integer :: stationIndex, phaseIndex, levelIndex
-    real    :: correctionValue
+    real(8) :: biasEstimate, correctionValue
     character(len=9) :: stationId
 
+    if (.not.initialized) call bcc_readConfig()
+    
     if ( aiRevOnly ) return
     
     nulcoeff = 0
-    ierr = fnom(nulcoeff, biasCorrectionFileName, 'FMT+R/O', 0)
+    ierr = fnom(nulcoeff, biasEstimateFile, 'FMT+R/O', 0)
     if ( ierr /= 0 ) then
-      call utl_abort('bcc_readAIBcor: unable to open airplanes bias correction file ' // biasCorrectionFileName )
+      call utl_abort('bcc_readAIBiases: unable to open airplanes bias correction file ' // biasEstimateFile )
     end if
     read (nulcoeff, '(i5)', iostat=ierr ) nbAircrafts
     if ( ierr /= 0 ) then
-      call utl_abort('bcc_readAIBcor: error 1 while reading airplanes bias correction file ' // biasCorrectionFileName )
+      call utl_abort('bcc_readAIBiases: error 1 while reading airplanes bias correction file ' // biasEstimateFile )
     end if
     
     allocate( ttCorrections(nAircraftMax,nPhases,nLevels) )
@@ -141,13 +146,17 @@ CONTAINS
     do stationIndex=1,nbAircrafts
       do phaseIndex=1,3
         do levelIndex=1,5
-          read (nulcoeff, *, iostat=ierr) stationId,correctionValue
+          read (nulcoeff, *, iostat=ierr) stationId, biasEstimate
           if ( ierr /= 0 ) then
-            call utl_abort('bcc_readAIBcor: error 2 while reading airplanes bias correction file ' // biasCorrectionFileName )
+            call utl_abort('bcc_readAIBiases: error 2 while reading airplanes bias correction file ' // biasEstimateFile )
           end if
-          if ( correctionValue == 99. ) correctionValue = MPC_missingValue_R8
+          if ( biasEstimate == 99.0D0 ) then
+            correctionValue = MPC_missingValue_R8
+          else
+            correctionValue = -1.0D0*biasEstimate
+          end if
           ttCorrections(stationIndex,phaseIndex,levelIndex) = correctionValue
-          aircraftIds(stationIndex)                       = stationId
+          aircraftIds(stationIndex)                         = stationId
           !print*, stationIndex, phaseIndex, levelIndex, aircraftIds(stationIndex), ttCorrections(stationIndex,phaseIndex,levelIndex)
         end do
       end do
@@ -156,17 +165,17 @@ CONTAINS
 
     ! Check for bulk bias corrections at start of file
     if ( aircraftIds(1) /= "BULKBCORS" ) then
-      call utl_abort('bcc_readAIBcor: ERROR: Bulk bias corrections are missing in bias correction file!' )
+      call utl_abort('bcc_readAIBiases: ERROR: Bulk bias corrections are missing in bias correction file!' )
     end if
 
-  end subroutine bcc_readAIBcor
+  end subroutine bcc_readAIBiases
 
   !-----------------------------------------------------------------------
   ! bcc_applyAIBcor
   !-----------------------------------------------------------------------
   subroutine bcc_applyAIBcor(obsSpaceData)
     !
-    ! :Purpose:  to fill OBS_BCOR column of ObsSpaceData body with aircraft TT bias correction 
+    ! :Purpose:  to apply aircraft (AI) bias corrections to observations in ObsSpaceData
     !
     implicit none
     !Arguments:
@@ -176,14 +185,17 @@ CONTAINS
     integer  :: flag, phase, bufrCode
     integer  :: phaseIndex, levelIndex, stationIndex, stationNumber
     integer  :: countTailCorrections,  countBulkCorrections
+    integer  :: headerFlag
     real(8)  :: corr, tt, oldCorr, pressure
     character(len=9) :: stnid, stnId1, stnId2
 
+    if (.not.initialized) call bcc_readConfig()
+    
     if ( .not. aiBiasActive ) return
 
     write(*,*) "bcc_applyAIBcor: start"
 
-    if ( .not.aiRevOnly ) call bcc_readAIBcor(aiBcFile)
+    if ( .not.aiRevOnly ) call bcc_readAIBiases(aiBcFile)
     
     countTailCorrections = 0
     countBulkCorrections = 0
@@ -192,6 +204,8 @@ CONTAINS
     HEADER: do
       headerIndex = obs_getHeaderIndex(obsSpaceData)
       if ( headerIndex < 0 ) exit HEADER
+      
+      headerFlag = obs_headElem_i(obsSpaceData, OBS_ST1, headerIndex )
       
       call obs_set_current_body_list(obsSpaceData, headerIndex)
 
@@ -210,13 +224,17 @@ CONTAINS
           oldCorr = obs_bodyElem_r(obsSpaceData, OBS_BCOR, bodyIndex )
           corr = MPC_missingValue_R8
           
-          if ( tt /= real(MPC_missingValue_R8,pre_obsReal) ) then
+          if ( tt /= MPC_missingValue_R8 ) then
           
-            if ( btest(flag, 6) .and. oldCorr /= real(MPC_missingValue_R8,pre_obsReal) ) then
-              tt = tt + oldCorr
+            if ( btest(flag, 6) .and. oldCorr /= MPC_missingValue_R8 ) then
+              if ( btest(headerFlag, 15) ) then
+                tt = tt - oldCorr
+              else
+                tt = tt + oldCorr
+              end if
               flag = ibclr(flag, 6)
             end if
-            if ( aiRevOnly ) corr = 0.0
+            if ( aiRevOnly ) corr = 0.0D0
              
             if ( .not.aiRevOnly ) then
                 
@@ -224,23 +242,23 @@ CONTAINS
 
               ! Get level index and current (mar 2020) static bulk corrections applied to AI TT data at derivate stage
               if ( (pressure <= 1100.d0) .and. (pressure > 700.d0) ) then
-                corr = 0.0
+                corr = 0.0D0
                 levelIndex = 5
               else if ( (pressure <= 700.d0)  .and. (pressure > 500.d0) ) then
-                corr = 0.1
+                corr = -0.1D0
                 levelIndex = 4
               else if ( (pressure <= 500.d0)  .and. (pressure > 400.d0) ) then
-                corr = 0.2
+                corr = -0.2D0
                 levelIndex = 3
               else if ( (pressure <= 400.d0)  .and. (pressure > 300.d0) ) then
-                corr = 0.3
+                corr = -0.3D0
                 levelIndex = 2
               else if ( (pressure <= 300.d0)  .and. (pressure > 100.d0) ) then
-                corr = 0.5
+                corr = -0.5D0
                 levelIndex = 1
               else 
                 levelIndex = 0
-                corr = 0.0
+                corr = 0.0D0
               end if
  
               codtyp = obs_headElem_i(obsSpaceData, OBS_ITY, headerIndex)
@@ -296,7 +314,7 @@ CONTAINS
               end if
             
               ! Apply the bias correction (bulk or new) and set the "bias corrected" bit in TT data flag ON
-              tt = tt - corr
+              tt = tt + corr
               flag = ibset(flag, 6)
             
             end if
@@ -310,6 +328,10 @@ CONTAINS
         end if
         
       end do BODY
+      
+      headerFlag = ibset(headerFlag, 15)
+      call obs_headSet_i( obsSpaceData, OBS_ST1, headerIndex, headerFlag )
+      
     end do HEADER
     
     if ( countBulkCorrections + countTailCorrections /= 0 ) then
@@ -327,32 +349,34 @@ CONTAINS
   end subroutine bcc_applyAIBcor
 
   !-----------------------------------------------------------------------
-  ! bcc_readGPBcor
+  ! bcc_readGPBiases
   !-----------------------------------------------------------------------
-  subroutine bcc_readGPBcor(biasCorrectionFileName)
+  subroutine bcc_readGPBiases(biasEstimateFile)
     !
-    ! :Purpose: Read GB-GPS bias corrections (mean ZTD O-A [mm] by station).
+    ! :Purpose: Read GB-GPS bias estimates (mean ZTD O-A [mm] by station) and fill bias correction array ztdCorrections.
     !           Missing value = -999.00
     !
     implicit none
     !Arguments:
-    character(len=*), intent(in) :: biasCorrectionFileName
+    character(len=*), intent(in) :: biasEstimateFile
     !Locals:
     integer :: ierr, nulcoeff
     integer :: stationIndex
-    real    :: correctionValue
+    real(8) :: biasEstimate
     character(len=9) :: stationId
 
+    if (.not.initialized) call bcc_readConfig()
+    
     if ( gpRevOnly ) return
     
     nulcoeff = 0
-    ierr = fnom(nulcoeff, biasCorrectionFileName, 'FMT+R/O', 0)
+    ierr = fnom(nulcoeff, biasEstimateFile, 'FMT+R/O', 0)
     if ( ierr /= 0 ) then
-      call utl_abort('bcc_readGPBcor: unable to open GB-GPS bias correction file ' // biasCorrectionFileName )
+      call utl_abort('bcc_readGPBiases: unable to open GB-GPS bias correction file ' // biasEstimateFile )
     end if
     read (nulcoeff, '(i5)', iostat=ierr ) nbGpStations
     if ( ierr /= 0 ) then
-      call utl_abort('bcc_readGPBcor: error 1 while reading GB-GPS bias correction file ' // biasCorrectionFileName )
+      call utl_abort('bcc_readGPBiases: error 1 while reading GB-GPS bias correction file ' // biasEstimateFile )
     end if
 
     allocate( ztdCorrections(nStationMaxGP) )
@@ -361,23 +385,23 @@ CONTAINS
     ztdCorrections(:) =  MPC_missingValue_R8
     
     do stationIndex=1,nbGpStations
-       read (nulcoeff, *, iostat=ierr) stationId,correctionValue
+       read (nulcoeff, *, iostat=ierr) stationId, biasEstimate
        if ( ierr /= 0 ) then
-          call utl_abort('bcc_readGPBcor: error 2 while reading GB-GPS bias correction file ' // biasCorrectionFileName )
+          call utl_abort('bcc_readGPBiases: error 2 while reading GB-GPS bias correction file ' // biasEstimateFile )
        end if
-       if ( correctionValue /= -999.00 ) ztdCorrections(stationIndex) = -correctionValue/1000.  ! mm to m
+       if ( biasEstimate /= MPC_missingValue_R8 ) ztdCorrections(stationIndex) = -1.0D0*(biasEstimate/1000.0D0)  ! mm to m
        gpsStations(stationIndex) = stationId
     end do
     ierr = fclos(nulcoeff)
     
-  end subroutine bcc_readGPBcor
+  end subroutine bcc_readGPBiases
 
   !-----------------------------------------------------------------------
   ! bcc_applyGPBcor
   !-----------------------------------------------------------------------
   subroutine bcc_applyGPBcor(obsSpaceData)
     !
-    ! :Purpose:  to fill OBS_BCOR column of ObsSpaceData body with GB-GPS ZTD bias correction 
+    ! :Purpose:  to apply GB-GPS (GP) ZTD bias corrections to ZTD observations in ObsSpaceData
     !
     implicit none
     !Arguments:
@@ -390,11 +414,13 @@ CONTAINS
     real(8)  :: corr, ztd, oldCorr
     character(len=9) :: stnid, stnId1, stnId2
 
+    if (.not.initialized) call bcc_readConfig()
+    
     if ( .not. gpBiasActive ) return
 
     write(*,*) "bcc_applyGPBcor: start"
 
-    if ( .not.gpRevOnly ) call bcc_readGPBcor(gpBcFile)
+    if ( .not.gpRevOnly ) call bcc_readGPBiases(gpBcFile)
     
     nbCorrected = 0
 
@@ -423,10 +449,10 @@ CONTAINS
           
           corr = MPC_missingValue_R8
           
-          if ( ztd /= real(MPC_missingValue_R8,pre_obsReal) ) then  
+          if ( ztd /= MPC_missingValue_R8 ) then  
             
             ! Remove any previous bias correction
-            if ( btest(flag, 6) .and. oldCorr /= real(MPC_missingValue_R8,pre_obsReal) ) then
+            if ( btest(flag, 6) .and. oldCorr /= MPC_missingValue_R8 ) then
               ztd = ztd - oldCorr
               flag = ibclr(flag, 6)
             end if
@@ -454,10 +480,10 @@ CONTAINS
                 nbCorrected = nbCorrected + 1
                 flag = ibset(flag, 6)
               else 
-                corr = 0.0
+                corr = 0.0D0
               end if
             else
-              corr = 0.0
+              corr = 0.0D0
             end if
 
           end if
@@ -495,7 +521,9 @@ CONTAINS
     implicit none
     character(len=*),intent(in) :: obsFam
 
-    bcc_biasActive = ( bcc_gpBiasActive .and. trim(obsFam) == 'GP')
+    if (.not.initialized) call bcc_readConfig()
+    
+    bcc_biasActive = (bcc_gpBiasActive .and. trim(obsFam) == 'GP') .or. (bcc_aiBiasActive .and. trim(obsFam) == 'AI')
 
   end function bcc_biasActive